Remove patroni support (discovery and metrics).
