+++ internal/pgscv/config_test.go
@@ -397,8 +397,6 @@ func Test_newConfigFromEnv(t *testing.T) {
 				"POSTGRES_DSN_EXAMPLE1":    "example_dsn",
 				"PGBOUNCER_DSN":            "example_dsn",
 				"PGBOUNCER_DSN_EXAMPLE2":   "example_dsn",
-				"PATRONI_URL":              "example_url",
-				"PATRONI_URL_EXAMPLE3":     "example_url",
 				"PGSCV_AUTH_USERNAME":      "user",
 				"PGSCV_AUTH_PASSWORD":      "pass",
 				"PGSCV_AUTH_KEYFILE":       "keyfile.key",
@@ -414,8 +412,6 @@ func Test_newConfigFromEnv(t *testing.T) {
 					"EXAMPLE1":  {ServiceType: model.ServiceTypePostgresql, Conninfo: "example_dsn"},
 					"pgbouncer": {ServiceType: model.ServiceTypePgbouncer, Conninfo: "example_dsn"},
 					"EXAMPLE2":  {ServiceType: model.ServiceTypePgbouncer, Conninfo: "example_dsn"},
-					"patroni":   {ServiceType: model.ServiceTypePatroni, BaseURL: "example_url"},
-					"EXAMPLE3":  {ServiceType: model.ServiceTypePatroni, BaseURL: "example_url"},
 				},
 				AuthConfig: http.AuthConfig{
 					Username: "user",
@@ -433,10 +429,6 @@ func Test_newConfigFromEnv(t *testing.T) {
 		{
 			valid:   false, // Invalid pgbouncer DSN key
 			envvars: map[string]string{"PGBOUNCER_DSN_": "example_dsn"},
-		},
-		{
-			valid:   false, // Invalid patroni URL key
-			envvars: map[string]string{"PATRONI_URL_": "example_dsn"},
 		},
 	}
